+++ gnsq/httpclient.py
@@ -1,24 +1,56 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
+import json
+
 import urllib3
 
-try:
-    import simplejson as json
-except ImportError:
-    import json  # pyflakes.ignore
-
-from .decorators import cached_property
 from .errors import NSQHttpError
+from .version import __version__
+
+USERAGENT = 'gnsq/{}'.format(__version__)
+
+
+def _encode(value):
+    if isinstance(value, bytes):
+        return value
+    return value.encode('utf-8')
+
+
+def _encode_dict(value):
+    if value is None:
+        return None
+    return {_encode(k): _encode(v) for k, v in value.items()}
 
 
 class HTTPClient(object):
-    @cached_property
-    def http(self):
-        return urllib3.connection_from_url(url=self.base_url)
+    @classmethod
+    def from_url(cls, url, **kwargs):
+        """Create a client from a url."""
+        url = urllib3.util.parse_url(url)
+        if url.host:
+            kwargs.setdefault('host', url.host)
 
-    def http_request(self, method, url, **kwargs):
-        response = self.http.request_encode_url(method, url, **kwargs)
+        if url.port:
+            kwargs.setdefault('port', url.port)
+
+        if url.scheme == 'https':
+            kwargs.setdefault('connection_class', urllib3.HTTPSConnectionPool)
+
+        return cls(**kwargs)
+
+    def __init__(self, host, port, useragent=USERAGENT,
+                 connection_class=urllib3.HTTPConnectionPool, **kwargs):
+        self.useragent = useragent
+        self._connection = connection_class(host, port, **kwargs)
+
+    def _request(self, method, url, headers={}, fields=None, **kwargs):
+        headers.setdefault('Accept', 'application/vnd.nsq version=1.0')
+        headers.setdefault('User-Agent', self.useragent)
+
+        response = self._connection.request_encode_url(
+            method, url, headers=_encode_dict(headers),
+            fields=_encode_dict(fields), **kwargs)
 
         if 'application/json' in response.getheader('content-type', ''):
             return self._http_check_json(response)
@@ -27,7 +59,7 @@ class HTTPClient(object):
 
     def _http_check(self, response):
         if response.status != 200:
-            raise NSQHttpError('http error <%s>' % response.status)
+            raise NSQHttpError('http error <{}>'.format(response.status))
         return response.data
 
     def _http_check_json(self, response):
@@ -38,16 +70,13 @@ class HTTPClient(object):
 
         if response.status != 200:
             status_txt = data.get('status_txt', 'http error')
-            raise NSQHttpError('%s <%s>' % (status_txt, response.status))
+            raise NSQHttpError('{} <{}>'.format(status_txt, response.status))
 
         # Handle 1.0.0-compat vs 0.x versions
-        if 'data' in data:
-            return data['data']
-        else:
+        try:
+            data['data']
+        except KeyError:
             return data
 
-    def http_get(self, url, **kwargs):
-        return self.http_request('GET', url, **kwargs)
-
-    def http_post(self, url, **kwargs):
-        return self.http_request('POST', url, **kwargs)
+    def __repr__(self):
+        return '<HTTPClient host={!r} port={!r}>'.format(self.host, self.port)